Added backend permission checks

+++ src/GraphQL/Mutations/PurgeElement.php
@@ -7,7 +7,10 @@
 
 namespace Aimeos\Cms\GraphQL\Mutations;
 
+use Illuminate\Support\Facades\Auth;
 use Aimeos\Cms\Models\Element;
+use Aimeos\Cms\Permission;
+use GraphQL\Error\Error;
 
 
 final class PurgeElement
@@ -18,6 +21,10 @@ final class PurgeElement
      */
     public function __invoke( $rootValue, array $args ) : array
     {
+        if( !Permission::can( 'element:purge', Auth::user() ) ) {
+            throw new Error( 'Insufficient permissions' );
+        }
+
         $items = Element::withTrashed()->whereIn( 'id', $args['id'] )->get();
         Element::whereIn( 'id', $items->pluck( 'id' ) )->forceDelete();
